# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Hakimi Healthcare API
  version: 1.0.0
  license:
    name: UNLICENSED
    identifier: UNLICENSED
  description: |
    REST API for Hakimi / ሀኪሜ, a healthcare appointment platform for Ethiopia.

    Notes:
    - Workforce API operations require an OIDC bearer access token with workforce MFA.
    - Authorization is default deny and derives roles and facility scope from server-side PostgreSQL state.
    - Patient authentication and self-service remain unavailable.
    - `DELETE` on facilities, practitioners, and assignments performs a soft delete.
    - Practitioner facility primary reassignment is handled atomically inside one transaction.
servers:
  - url: http://127.0.0.1:3001
security:
  - workforceBearer: []
...
paths:
...
  /api/v1/facilities:
    post:
      tags: [Facilities]
      summary: Create a healthcare facility
      operationId: createHealthcareFacility
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHealthcareFacilityInput'
            examples:
              clinic:
                summary: Fictional clinic
                value:
                  code: ADD-CLINIC-001
                  name: Addis Sunrise Clinic
                  facilityType: clinic
                  licenseNumber: LIC-ADD-001
                  phone: '+251911111111'
                  email: contact@addissunrise.example
                  region: Addis Ababa
                  city: Addis Ababa
                  addressLine: Bole Road
                  isActive: true
      responses:
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

        '201':
          description: Created facility
          headers:
            Location:
              description: URL of the created facility
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthcareFacility'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/FacilityConflict'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Facilities]
      summary: List healthcare facilities
      operationId: listHealthcareFacilities
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: facilityType
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/FacilityType'
        - name: region
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - name: city
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - name: isActive
          in: query
          required: false
          schema:
            type: boolean
        - name: search
          in: query
          required: false
          schema:
            type: string
            minLength: 2
            maxLength: 100
      responses:
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

        '200':
          description: Paged list of facilities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthcareFacilityListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/facilities/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Facilities]
      summary: Get a facility by id
      operationId: getHealthcareFacilityById
      responses:
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

        '200':
          description: Facility found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthcareFacility'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/FacilityNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Facilities]
      summary: Update a facility
      operationId: updateHealthcareFacility
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateHealthcareFacilityInput'
            examples:
              rename:
                summary: Rename a facility
                value:
                  name: Addis Sunrise Clinic Updated
      responses:
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

        '200':
          description: Updated facility
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthcareFacility'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/FacilityNotFound'
        '409':
          $ref: '#/components/responses/FacilityConflict'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Facilities]
      summary: Deactivate a facility
      description: |
        Soft delete only. This operation sets `isActive` to `false` and updates `updatedAt`.
        The record remains available through `GET /api/v1/facilities/{id}`.
      operationId: deactivateHealthcareFacility
      responses:
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

        '204':
          description: Facility deactivated or already inactive
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/FacilityNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/practitioners/{practitionerId}/facilities:
    parameters:
      - name: practitionerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Assignments]
      summary: Create a practitioner-facility assignment
      description: |
        Creates an assignment for the practitioner in the URL.
        When `isPrimary=true` and `isActive=true`, the API clears the prior active primary assignment atomically.
      operationId: createPractitionerAssignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePractitionerAssignmentInput'
            examples:
              primary:
                summary: Primary clinic assignment
                value:
                  facilityId: 33333333-3333-4333-8333-333333333333
                  roleTitle: Physician
                  department: Internal Medicine
                  isPrimary: true
                  isActive: true
      responses:
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

        '201':
          description: Created assignment
          headers:
            Location:
              description: URL of the created assignment
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PractitionerFacilityAssignment'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/AssignmentNotFound'
        '409':
          $ref: '#/components/responses/AssignmentConflict'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Assignments]
      summary: List practitioner assignments
      operationId: listPractitionerAssignments
      responses:
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

        '200':
          description: Assignment list including active and inactive records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PractitionerAssignmentListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/PractitionerNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/practitioners/{practitionerId}/facilities/{assignmentId}:
    parameters:
      - name: practitionerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: assignmentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [Assignments]
      summary: Update a practitioner-facility assignment
      description: |
        `facilityId` cannot be changed.
        If an assignment becomes inactive, the API clears `isPrimary`.
      operationId: updatePractitionerAssignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePractitionerAssignmentInput'
            examples:
              promote:
                summary: Promote the assignment to primary
                value:
                  isPrimary: true
      responses:
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

        '200':
          description: Updated assignment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PractitionerFacilityAssignment'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/AssignmentNotFound'
        '409':
          $ref: '#/components/responses/AssignmentConflict'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Assignments]
      summary: Deactivate a practitioner-facility assignment
      description: |
        Soft delete only. This operation sets `isActive` to `false`, clears `isPrimary`, and updates `updatedAt`.
        Repeating the delete against the same assignment remains idempotent.
      operationId: deactivatePractitionerAssignment
      responses:
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

        '204':
          description: Assignment deactivated or already inactive
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/AssignmentNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    workforceBearer:
      type: oauth2
      description: |
        Workforce-only OIDC Authorization Code with PKCE. The resource server
        requires a short-lived signed JWT access token, approved workforce MFA,
        and an active server-provisioned actor, role, scope, and local session.
        Patient accounts and role-administration APIs are not available.
      flows:
        authorizationCode:
          authorizationUrl: https://identity.example.com/oauth2/authorize
          tokenUrl: https://identity.example.com/oauth2/token
          scopes: {}
  schemas:
    FacilityType:
      type: string
      enum:
        - hospital
        - clinic
        - health_center
        - diagnostic_center
        - pharmacy
    HealthcareFacility:
      type: object
      additionalProperties: false
      required:
        - id
        - code
        - name
        - facilityType
        - licenseNumber
        - phone
        - email
        - region
        - city
        - addressLine
        - isActive
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        facilityType:
          $ref: '#/components/schemas/FacilityType'
        licenseNumber:
          type:
            - string
            - 'null'
        phone:
          type:
            - string
            - 'null'
        email:
          type:
            - string
            - 'null'
        region:
          type: string
        city:
          type: string
        addressLine:
          type:
            - string
            - 'null'
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateHealthcareFacilityInput:
      type: object
      additionalProperties: false
      required:
        - code
        - name
        - facilityType
        - region
        - city
      properties:
        code:
          type: string
          minLength: 2
          maxLength: 50
        name:
          type: string
          minLength: 2
          maxLength: 200
        facilityType:
          $ref: '#/components/schemas/FacilityType'
        licenseNumber:
          type:
            - string
            - 'null'
          maxLength: 100
        phone:
          type:
            - string
            - 'null'
          maxLength: 30
        email:
          type:
            - string
            - 'null'
          format: email
          maxLength: 254
        region:
          type: string
          minLength: 2
          maxLength: 100
        city:
          type: string
          minLength: 2
          maxLength: 100
        addressLine:
          type:
            - string
            - 'null'
          maxLength: 300
        isActive:
          type: boolean
          default: true
    UpdateHealthcareFacilityInput:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        code:
          type: string
          minLength: 2
          maxLength: 50
        name:
          type: string
          minLength: 2
          maxLength: 200
        facilityType:
          $ref: '#/components/schemas/FacilityType'
        licenseNumber:
          type:
            - string
            - 'null'
          maxLength: 100
        phone:
          type:
            - string
            - 'null'
          maxLength: 30
        email:
          type:
            - string
            - 'null'
          format: email
          maxLength: 254
        region:
          type: string
          minLength: 2
          maxLength: 100
        city:
          type: string
          minLength: 2
          maxLength: 100
        addressLine:
          type:
            - string
            - 'null'
          maxLength: 300
        isActive:
          type: boolean
    HealthcareFacilityPagination:
      type: object
      additionalProperties: false
      required:
        - page
        - pageSize
        - totalItems
        - totalPages
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
        totalItems:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    HealthcareFacilityListResponse:
      type: object
      additionalProperties: false
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/HealthcareFacility'
        pagination:
          $ref: '#/components/schemas/HealthcareFacilityPagination'
...
    PractitionerAssignmentFacilitySummary:
      type: object
      additionalProperties: false
      required:
        - id
        - code
        - name
        - facilityType
        - region
        - city
        - isActive
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        facilityType:
          $ref: '#/components/schemas/FacilityType'
        region:
          type: string
        city:
          type: string
        isActive:
          type: boolean
    PractitionerFacilityAssignment:
      type: object
      additionalProperties: false
      required:
        - id
        - practitionerId
        - facilityId
        - roleTitle
        - department
        - isPrimary
        - isActive
        - createdAt
        - updatedAt
        - facility
      properties:
        id:
          type: string
          format: uuid
        practitionerId:
          type: string
          format: uuid
        facilityId:
          type: string
          format: uuid
        roleTitle:
          type: string
        department:
          type:
            - string
            - 'null'
        isPrimary:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        facility:
          $ref: '#/components/schemas/PractitionerAssignmentFacilitySummary'
    CreatePractitionerAssignmentInput:
      type: object
      additionalProperties: false
      required:
        - facilityId
        - roleTitle
      properties:
        facilityId:
          type: string
          format: uuid
        roleTitle:
          type: string
          minLength: 2
          maxLength: 100
        department:
          type:
            - string
            - 'null'
          maxLength: 100
        isPrimary:
          type: boolean
          default: false
        isActive:
          type: boolean
          default: true
    UpdatePractitionerAssignmentInput:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        roleTitle:
          type: string
          minLength: 2
          maxLength: 100
        department:
          type:
            - string
            - 'null'
          maxLength: 100
        isPrimary:
          type: boolean
        isActive:
          type: boolean
    PractitionerAssignmentListResponse:
      type: object
      additionalProperties: false
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PractitionerFacilityAssignment'
...
    ApiErrorDetail:
      type: object
      additionalProperties: false
      required:
        - field
        - message
      properties:
        field:
          type: string
        message:
          type: string
    ApiErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - INVALID_JSON
                - AUTHENTICATION_REQUIRED
                - FORBIDDEN
                - FACILITY_NOT_FOUND
                - FACILITY_CODE_CONFLICT
                - FACILITY_LICENSE_CONFLICT
                - FACILITY_INACTIVE
                - PRACTITIONER_NOT_FOUND
                - PRACTITIONER_CODE_CONFLICT
                - PRACTITIONER_LICENSE_CONFLICT
                - INACTIVE_PRACTITIONER
                - INACTIVE_FACILITY
                - ASSIGNMENT_NOT_FOUND
                - ASSIGNMENT_CONFLICT
                - PATIENT_NOT_FOUND
                - INACTIVE_PATIENT
                - PATIENT_REGISTRATION_NOT_FOUND
                - PATIENT_REGISTRATION_CONFLICT
                - APPOINTMENT_NOT_FOUND
                - APPOINTMENT_CONFLICT
                - APPOINTMENT_STATE_CONFLICT
                - INTERNAL_ERROR
            message:
              type: string
            details:
              type: array
              items:
                $ref: '#/components/schemas/ApiErrorDetail'
  responses:
    AuthenticationRequired:
      description: Missing, invalid, expired, revoked, or otherwise ineligible workforce authentication
      headers:
        WWW-Authenticate:
          description: Bearer authentication challenge
          schema:
            type: string
            const: Bearer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            authenticationRequired:
              value:
                error:
                  code: AUTHENTICATION_REQUIRED
                  message: Authentication required
    Forbidden:
      description: Authenticated workforce actor is not permitted to perform this operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            forbidden:
              value:
                error:
                  code: FORBIDDEN
                  message: Forbidden
    ValidationError:
      description: Validation failure or malformed request body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            validation:
              value:
                error:
                  code: VALIDATION_ERROR
                  message: Request validation failed
                  details:
                    - field: name
                      message: Name is required
    FacilityNotFound:
      description: Facility not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            notFound:
              value:
                error:
                  code: FACILITY_NOT_FOUND
                  message: Facility not found
    FacilityConflict:
      description: Facility code or license number conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            codeConflict:
              value:
                error:
                  code: FACILITY_CODE_CONFLICT
                  message: Facility code already exists
            licenseConflict:
              value:
                error:
                  code: FACILITY_LICENSE_CONFLICT
                  message: License number already exists
    PractitionerNotFound:
      description: Practitioner not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            notFound:
              value:
                error:
                  code: PRACTITIONER_NOT_FOUND
                  message: Practitioner not found
...
    AssignmentNotFound:
      description: Assignment not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            notFound:
              value:
                error:
                  code: ASSIGNMENT_NOT_FOUND
                  message: Assignment not found
    AssignmentConflict:
      description: Assignment conflict or duplicate assignment
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            duplicate:
              value:
                error:
                  code: ASSIGNMENT_CONFLICT
                  message: Assignment already exists
            inactivePractitioner:
              value:
                error:
                  code: INACTIVE_PRACTITIONER
                  message: Practitioner is inactive
            inactiveFacility:
              value:
                error:
                  code: INACTIVE_FACILITY
                  message: Facility is inactive
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            internal:
              value:
                error:
                  code: INTERNAL_ERROR
                  message: Internal server error